backend/requirements.txt
fastapi
uvicorn
firebase-admin
pydantic
python-dotenv
easyocr
Pillow
python-multipart
torch
torchvision
